Refactor package configuration to consolidate defaults into pkg.py
Move container and default configurations from separate files into the
main pkg.py module for each package. This simplifies the package
structure and ensures defaults are properly maintained without being
reset by package instantiation. Removes intermediate config files
(container.py and default.py) and updates core package handling logic.
