+++ builtin/builtin/vpic/pkg.py
@@ -2,22 +2,33 @@
 VPIC-Kokkos — Vector Particle-In-Cell plasma physics simulation.
 GPU-accelerated, relativistic, kinetic PIC code from Los Alamos National Lab.
 """
-from jarvis_cd.core.route_pkg import RouteApp
+from jarvis_cd.core.pkg import Application
+from jarvis_cd.shell import Exec, MpiExecInfo, PsshExecInfo
+from jarvis_cd.shell.process import Mkdir, Rm
+import os
 
 
-class Vpic(RouteApp):
+class Vpic(Application):
     """
-    Router class for VPIC deployment.
+    Merged VPIC class supporting both default (bare-metal) and container deployment.
+
+    Set deploy_mode='container' to build and run VPIC inside a Docker/Podman/Apptainer
+    container with Kokkos CUDA backend.  Set deploy_mode='default' to use a
+    system-installed vpic binary via MPI.
     """
+
+    def _init(self):
+        self.deck_binary = None
 
     def _configure_menu(self):
-        base_menu = super()._configure_menu()
-        for item in base_menu:
-            if item['name'] == 'deploy_mode':
-                item['choices'] = ['default', 'container']
-                break
-
-        return base_menu + [
+        return [
+            {
+                'name': 'deploy_mode',
+                'msg': 'Deployment mode',
+                'type': str,
+                'choices': ['default', 'container'],
+                'default': 'default',
+            },
             {
                 'name': 'nprocs',
                 'msg': 'Number of MPI processes',
@@ -62,3 +73,154 @@ class Vpic(RouteApp):
                 'default': 'sci-hpc-base',
             },
         ]
+
+    # ------------------------------------------------------------------
+    # Container Dockerfile generators
+    # ------------------------------------------------------------------
+
+    def _build_phase(self) -> str:
+        """
+        Return the BUILD container Dockerfile, or None when not in container mode.
+        """
+        if self.config.get('deploy_mode') != 'container':
+            return None
+        cuda_arch = self.config.get('cuda_arch', 80)
+        base = self.config.get('base_image', 'sci-hpc-base')
+        return f"""FROM {base}
+
+ARG CUDA_ARCH={cuda_arch}
+
+# Clone VPIC-Kokkos with bundled Kokkos (cached until repo changes)
+RUN git clone --recursive https://github.com/lanl/vpic-kokkos.git /opt/vpic-kokkos
+
+# Build VPIC core library with Kokkos CUDA
+RUN cd /opt/vpic-kokkos \\
+    && cmake -S . -B build \\
+        -DCMAKE_BUILD_TYPE=Release \\
+        -DENABLE_KOKKOS_CUDA=ON \\
+        -DBUILD_INTERNAL_KOKKOS=ON \\
+        "-DKokkos_ARCH_AMPERE${{CUDA_ARCH}}=ON" \\
+        -DCMAKE_CXX_COMPILER="$(pwd)/kokkos/bin/nvcc_wrapper" \\
+    && cmake --build build -j$(nproc)
+
+# Patch deck-compiler to link against CUDA stub library
+RUN sed -i \\
+    's|-lkokkossimd|-lkokkossimd -L/usr/local/cuda/lib64/stubs -lcuda|' \\
+    /opt/vpic-kokkos/build/bin/vpic
+
+ENV NVCC_WRAPPER_DEFAULT_COMPILER=g++
+ENV PATH=/opt/vpic-kokkos/build/bin:${{PATH}}
+"""
+
+    def _build_deploy_phase(self) -> str:
+        """
+        Return the DEPLOY container Dockerfile, or None when not in container mode.
+        """
+        if self.config.get('deploy_mode') != 'container':
+            return None
+        base = self.config.get('base_image', 'sci-hpc-base')
+        return f"""FROM {self.build_image_name} AS builder
+FROM {base}
+
+# Copy VPIC binaries and samples from build container
+COPY --from=builder /opt/vpic-kokkos/build/bin /opt/vpic-kokkos/build/bin
+COPY --from=builder /opt/vpic-kokkos/sample /opt/vpic-kokkos/sample
+COPY --from=builder /opt/vpic-kokkos/kokkos/bin/nvcc_wrapper /opt/vpic-kokkos/kokkos/bin/nvcc_wrapper
+
+ENV NVCC_WRAPPER_DEFAULT_COMPILER=g++
+ENV PATH=/opt/vpic-kokkos/build/bin:${{PATH}}
+
+CMD ["/bin/bash"]
+"""
+
+    # ------------------------------------------------------------------
+    # Configuration
+    # ------------------------------------------------------------------
+
+    def _configure(self, **kwargs):
+        """
+        Configure VPIC.
+
+        Calls super()._configure() which updates self.config and (when
+        deploy_mode == 'container') triggers build_phase / build_deploy_phase.
+
+        In default mode, also creates the run directory on all nodes.
+        """
+        super()._configure(**kwargs)
+
+        if self.config.get('deploy_mode') == 'default':
+            Mkdir(self.config['run_dir'],
+                  PsshExecInfo(hostfile=self.hostfile, env=self.env)).run()
+
+    # ------------------------------------------------------------------
+    # Lifecycle
+    # ------------------------------------------------------------------
+
+    def start(self):
+        """
+        Launch VPIC.
+
+        Branches on deploy_mode: compiles and runs inside the container for
+        container mode, or uses system-installed vpic binary for default mode.
+        """
+        run_dir = self.config['run_dir']
+
+        if self.config.get('deploy_mode') == 'container':
+            Mkdir(run_dir).run()
+
+            if self.config.get('deck'):
+                deck_file = self.config['deck']
+            else:
+                sample = self.config.get('sample_deck', 'harris')
+                deck_file = f'/opt/vpic-kokkos/sample/{sample}/{sample}.cxx'
+
+            deck_name = os.path.basename(deck_file).replace('.cxx', '')
+            nprocs = self.config.get('nprocs', 4)
+
+            # Step 1: Compile deck inside container
+            compile_cmd = (
+                f'bash -c "cp {deck_file} {run_dir}/ && cd {run_dir} && '
+                f'/opt/vpic-kokkos/build/bin/vpic {deck_name}.cxx"'
+            )
+            Exec(compile_cmd, self.container_exec_info(gpu=True)).run()
+
+            # Step 2: Run compiled binary inside container
+            run_cmd = (
+                f'bash -c "cd {run_dir} && mpirun --allow-run-as-root -n {nprocs} '
+                f'{run_dir}/{deck_name}.Linux"'
+            )
+            Exec(run_cmd, self.container_exec_info(gpu=True)).run()
+        else:
+            if self.config.get('deck'):
+                deck_file = self.config['deck']
+            else:
+                sample = self.config.get('sample_deck', 'harris')
+                deck_file = f'/opt/vpic-kokkos/sample/{sample}/{sample}.cxx'
+
+            # Step 1: Compile the deck
+            deck_name = os.path.basename(deck_file).replace('.cxx', '')
+            Exec(
+                f'cp {deck_file} {run_dir}/ && cd {run_dir} && vpic {deck_name}.cxx',
+                MpiExecInfo(nprocs=1, hostfile=self.hostfile, env=self.mod_env,
+                            cwd=run_dir)
+            ).run()
+
+            # Step 2: Run compiled binary
+            binary = f'{run_dir}/{deck_name}.Linux'
+            Exec(
+                binary,
+                MpiExecInfo(nprocs=self.config['nprocs'],
+                            ppn=self.config['ppn'],
+                            hostfile=self.hostfile,
+                            env=self.mod_env,
+                            cwd=run_dir)
+            ).run()
+
+    def stop(self):
+        """Stop VPIC (no-op — VPIC runs to completion)."""
+        pass
+
+    def clean(self):
+        """Remove VPIC run directory."""
+        Rm(self.config['run_dir'] + '*',
+           PsshExecInfo(hostfile=self.hostfile, env=self.env)).run()
