+++ jarvis_cd/core/container_pkg.py
@@ -1,233 +1,9 @@
 """
-Base classes for containerized application deployment.
+Deprecated: Container support is now built into Application (jarvis_cd/core/pkg.py).
+This module is kept for backwards compatibility only.
 """
-from .pkg import Application
-from ..shell import Exec, LocalExecInfo
-from pathlib import Path
+from .pkg import Application, Service
 
-
-class ContainerApplication(Application):
-    """
-    Base class for containerized application deployment using Docker/Podman/Apptainer.
-
-    Implements a two-phase container build strategy:
-
-    1. Build phase: A stable, cached container that compiles the application.
-       Named automatically as 'jarvis-build-{pkg_name}'.
-       Built once and reused across pipelines via Docker layer cache.
-
-    2. Deploy phase: A lightweight container that copies binaries from the build
-       container. Named after the pipeline. For apptainer, the deploy container
-       is additionally converted to a SIF file.
-
-    Subclasses must implement:
-      - _build_phase(): Return Dockerfile content for the build container
-      - _build_deploy_phase(): Return Dockerfile content for the deploy container
-    """
-
-    def _init(self):
-        """Initialize paths"""
-        pass
-
-    @property
-    def build_image_name(self) -> str:
-        """
-        Stable name for the BUILD container image.
-        Independent of pipeline name so it can be reused across pipelines.
-        """
-        pkg_name = self.pkg_type.split('.')[-1].replace('_', '-')
-        return f'jarvis-build-{pkg_name}'
-
-    @property
-    def deploy_image_name(self) -> str:
-        """Name for the DEPLOY container image (pipeline-specific)."""
-        if hasattr(self, 'pipeline') and self.pipeline:
-            return self.pipeline.name
-        return 'jarvis-deploy'
-
-    @property
-    def _container_engine(self) -> str:
-        """Get the container engine from pipeline config."""
-        return getattr(self.pipeline, 'container_engine', 'docker').lower()
-
-    @property
-    def _build_engine(self) -> str:
-        """Get the engine to use for building (apptainer needs docker/podman as intermediate)."""
-        engine = self._container_engine
-        if engine == 'apptainer':
-            import shutil
-            if shutil.which('docker'):
-                return 'docker'
-            elif shutil.which('podman'):
-                return 'podman'
-            else:
-                raise RuntimeError(
-                    "Apptainer requires docker or podman for the build phase. "
-                    "Neither was found in PATH."
-                )
-        return engine
-
-    def _build_phase(self) -> str:
-        """
-        Return the Dockerfile content for the BUILD container.
-
-        The build container is a heavy container that compiles the application
-        from source. It is built once with a stable name and cached between
-        pipeline runs via Docker layer cache.
-
-        Override this method in subclasses to provide application-specific
-        build steps. Return None to skip the build phase.
-
-        :return: Dockerfile content string, or None to skip
-        """
-        return None
-
-    def build_phase(self):
-        """
-        Build the BUILD container image and save Dockerfile to private_dir.
-
-        Uses the container engine's layer cache so repeated builds are fast
-        when the Dockerfile content hasn't changed. The image is named
-        '{build_image_name}' (stable, pipeline-independent).
-
-        :return: None
-        """
-        dockerfile_content = self._build_phase()
-        if not dockerfile_content:
-            return
-
-        # Save Dockerfile to private_dir for audit/reference
-        private_path = Path(self.private_dir)
-        private_path.mkdir(parents=True, exist_ok=True)
-        dockerfile_path = private_path / 'build.Dockerfile'
-
-        with open(dockerfile_path, 'w') as f:
-            f.write(dockerfile_content)
-
-        print(f"Building build container: {self.build_image_name}")
-        engine = self._build_engine
-        build_cmd = (
-            f"{engine} build -t {self.build_image_name} "
-            f"-f {dockerfile_path} {private_path}"
-        )
-        Exec(build_cmd, LocalExecInfo()).run()
-        print(f"Build container ready: {self.build_image_name}")
-
-    def _build_deploy_phase(self) -> str:
-        """
-        Return the Dockerfile content for the DEPLOY container.
-
-        The deploy container copies compiled binaries from the build container
-        into a leaner runtime image. It is named after the pipeline.
-
-        For apptainer, this docker image is additionally converted to a SIF file.
-
-        Override this method in subclasses. Return None to skip.
-
-        :return: Dockerfile content string, or None to skip
-        """
-        return None
-
-    def build_deploy_phase(self):
-        """
-        Build the DEPLOY container image and save Dockerfile to private_dir.
-
-        For docker/podman: builds image named after the pipeline.
-        For apptainer: builds docker image then converts to SIF file stored in
-        private_dir/{deploy_image_name}.sif.
-
-        :return: None
-        """
-        dockerfile_content = self._build_deploy_phase()
-        if not dockerfile_content:
-            return
-
-        private_path = Path(self.private_dir)
-        private_path.mkdir(parents=True, exist_ok=True)
-        dockerfile_path = private_path / 'deploy.Dockerfile'
-
-        with open(dockerfile_path, 'w') as f:
-            f.write(dockerfile_content)
-
-        engine = self._container_engine
-        build_engine = self._build_engine
-
-        print(f"Building deploy container: {self.deploy_image_name}")
-        build_cmd = (
-            f"{build_engine} build -t {self.deploy_image_name} "
-            f"-f {dockerfile_path} {private_path}"
-        )
-        Exec(build_cmd, LocalExecInfo()).run()
-
-        if engine == 'apptainer':
-            sif_path = private_path / f'{self.deploy_image_name}.sif'
-            print(f"Converting to Apptainer SIF: {sif_path}")
-            from ..shell.container_compose_exec import ApptainerBuildExec
-            apptainer_exec = ApptainerBuildExec(
-                self.deploy_image_name,
-                str(sif_path),
-                LocalExecInfo(),
-                source='docker-daemon'
-            )
-            apptainer_exec.run()
-            print(f"Apptainer SIF ready: {sif_path}")
-        else:
-            print(f"Deploy container ready: {self.deploy_image_name}")
-
-    def _configure(self, **kwargs):
-        """
-        Configure container deployment.
-        Triggers build and deploy phase container builds.
-
-        :param kwargs: Configuration parameters for this pkg.
-        :return: None
-        """
-        super()._configure(**kwargs)
-        self.build_phase()
-        self.build_deploy_phase()
-
-    def container_exec_info(self, gpu: bool = False) -> 'LocalExecInfo':
-        """
-        Return a LocalExecInfo pre-configured with container engine settings.
-
-        The returned ExecInfo, when passed to Exec, will cause the command to
-        be automatically wrapped for the appropriate container engine via
-        Exec._wrap_container_cmd().
-
-        For apptainer, container_image is set to the SIF file path under
-        private_dir. For docker/podman, it is set to the deploy image name.
-
-        :param gpu: Enable GPU passthrough into the container.
-        :return: LocalExecInfo with container, gpu, and container_image set.
-        """
-        from pathlib import Path
-        engine = self._container_engine
-        if engine == 'apptainer':
-            image = str(Path(self.private_dir) / f'{self.deploy_image_name}.sif')
-        else:
-            image = self.deploy_image_name
-        return LocalExecInfo(container=engine, gpu=gpu, container_image=image,
-                             env=self.mod_env)
-
-    def start(self):
-        """Start is handled at the pipeline level for containerized applications."""
-        pass
-
-    def stop(self):
-        """Stop is handled at the pipeline level for containerized applications."""
-        pass
-
-    def clean(self):
-        """Clean is handled at the pipeline level for containerized applications."""
-        pass
-
-    def kill(self):
-        """Kill is handled at the pipeline level for containerized applications."""
-        pass
-
-
-class ContainerService(ContainerApplication):
-    """
-    Alias for ContainerApplication following service naming conventions.
-    """
-    pass
+# Aliases kept for any external code that might import these
+ContainerApplication = Application
+ContainerService = Service
